Implemented CodeBytes MiniMax - worked on board class.
TODO: Board gen needs to happen in board class - can be a separate method. This prevents IO Exceptions and will make the generator easier to use with a board.
ALTERNATIVELY: Make the board class take a board object (strange) or return from board gen a 2D Array (better idea). This is because the board constructor expects a 2D since the board doesn't actually fill itself, but instead takes a pre-generated array from the other class.

Summation: CodeBytes will be easy to implement - it should be implemented already, in all actuality. Just need to fix the board constructor/boardgen and it should chug away.