Alpha-Beta is working (hopefully) now. Super fast compared to MiniMax.
ISSUE: Sometimes when the algorithm has a fantastic location (or even a decent location) to place a move, it completely ignored it and starts placing pieces in the far right column. No idea why. Alpha-Beta does it too, but the base code came from the original MiniMax algorithm, so it could have something to do with the evaluation mechanic.

TODO: Pick another algorithm, toss it in, and make a presentation.